refactor: move createVideoThumbNail method which needs application context to data_source (emojiDataSource)
