+++ android/app/src/main/java/com/goliath/emojihub/data_sources/DataSourceModule.kt
@@ -2,6 +2,8 @@ package com.goliath.emojihub.data_sources
 
 import com.goliath.emojihub.data_sources.local.X3dDataSource
 import com.goliath.emojihub.data_sources.local.X3dDataSourceImpl
+import com.goliath.emojihub.data_sources.remote.EmojiDataSource
+import com.goliath.emojihub.data_sources.remote.EmojiDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -12,6 +14,9 @@ import dagger.hilt.components.SingletonComponent
 abstract class DataSourceModule {
     @Binds
     abstract fun bindsX3dDataSource(impl: X3dDataSourceImpl): X3dDataSource
+
+    @Binds
+    abstract fun bindsEmojiDataSource(impl: EmojiDataSourceImpl): EmojiDataSource
 
     @Binds
     abstract fun bindsApiErrorController(impl: ApiErrorControllerImpl): ApiErrorController
