+++ android/app/src/main/java/com/goliath/emojihub/repositories/remote/EmojiRepository.kt
@@ -1,8 +1,5 @@
 package com.goliath.emojihub.repositories.remote
 
-import android.content.Context
-import android.graphics.Bitmap
-import android.media.MediaMetadataRetriever
 import android.util.Log
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
@@ -10,20 +7,17 @@ import androidx.paging.PagingData
 import com.goliath.emojihub.data_sources.EmojiFetchType
 import com.goliath.emojihub.data_sources.EmojiPagingSource
 import com.goliath.emojihub.data_sources.api.EmojiApi
+import com.goliath.emojihub.data_sources.remote.EmojiDataSource
 import com.goliath.emojihub.models.EmojiDto
 import com.goliath.emojihub.models.UploadEmojiDto
 import com.google.gson.Gson
-import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.Flow
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.asRequestBody
 import okhttp3.RequestBody.Companion.toRequestBody
-import retrofit2.HttpException
 import retrofit2.Response
 import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -41,7 +35,7 @@ interface EmojiRepository {
 @Singleton
 class EmojiRepositoryImpl @Inject constructor(
     private val emojiApi: EmojiApi,
-    @ApplicationContext private val context: Context
+    private val emojiDataSource: EmojiDataSource
 ): EmojiRepository {
     override suspend fun fetchEmojiList(sortByDate: Int): Flow<PagingData<EmojiDto>> {
         return Pager(
@@ -75,86 +69,43 @@ class EmojiRepositoryImpl @Inject constructor(
         val videoFileRequestBody = videoFile.asRequestBody("video/mp4".toMediaTypeOrNull())
         val videoFileMultipartBody = MultipartBody.Part.createFormData("file", videoFile.name, videoFileRequestBody)
 
-        val thumbnailFile = createVideoThumbnail(context, videoFile)
+        val thumbnailFile = emojiDataSource.createVideoThumbNail(videoFile)
         val thumbnailRequestBody = thumbnailFile!!
             .asRequestBody("image/jpg".toMediaTypeOrNull())
         val thumbnailMultipartBody = MultipartBody.Part.createFormData("thumbnail",
             thumbnailFile.name, thumbnailRequestBody)
 
-        return try {
-            emojiApi.uploadEmoji(videoFileMultipartBody, thumbnailMultipartBody, emojiDtoRequestBody)
-            true
-        }
-        catch (e: IOException) {
-            Log.e("EmojiRepository", "IOException")
-            false
-        }
-        catch (e: HttpException) {
-            Log.e("EmojiRepository", "HttpException")
-            false
-        }
-        catch (e: Exception) {
-            Log.e("EmojiRepository", "Unknown Exception: ${e.message}")
-            false
-        }
+        return emojiApi.uploadEmoji(videoFileMultipartBody, thumbnailMultipartBody, emojiDtoRequestBody)
+            .isSuccessful
     }
 
     override suspend fun saveEmoji(id: String): Result<Unit> {
-        return try {
-            val response = emojiApi.saveEmoji(id)
-            Log.d("EmojiRepository", "SaveEmoji Api response : ${response.code()}")
+        val response = emojiApi.saveEmoji(id)
+        Log.d("EmojiRepository", "SaveEmoji Api response : ${response.code()}")
 
-            if (response.isSuccessful) {
-                Log.d("EmojiRepository", "Successfully saved Emoji (Id: $id)")
-                Result.success(Unit)
-            } else {
-                Log.d("EmojiRepository", "Failed to save Emoji (Id: $id), ${response.code()}")
-                Result.failure(Exception("Failed to save Emoji (Id: $id), ${response.code()}"))
-            }
-        } catch (e: Exception) {
-            Result.failure(e)
+        return if (response.isSuccessful) {
+            Log.d("EmojiRepository", "Successfully saved Emoji (Id: $id)")
+            Result.success(Unit)
+        } else {
+            Log.d("EmojiRepository", "Failed to save Emoji (Id: $id), ${response.code()}")
+            Result.failure(Exception("Failed to save Emoji (Id: $id), ${response.code()}"))
         }
     }
 
     override suspend fun unSaveEmoji(id: String): Result<Unit> {
-        return try {
-            val response = emojiApi.unSaveEmoji(id)
-            Log.d("EmojiRepository", "UnSaveEmoji Api response : ${response.code()}")
-            if (response.isSuccessful) {
-                Log.d("EmojiRepository", "Successfully unsaved Emoji (Id: $id)")
-                Result.success(Unit)
-            } else {
-                Log.d("EmojiRepository", "Failed to unsave Emoji (Id: $id), ${response.code()}")
-                Result.failure(Exception("Failed to unsave Emoji (Id: $id), ${response.code()}"))
-            }
-        } catch (e: Exception) {
-            Result.failure(e)
+        val response = emojiApi.unSaveEmoji(id)
+        Log.d("EmojiRepository", "UnSaveEmoji Api response : ${response.code()}")
+
+        return if (response.isSuccessful) {
+            Log.d("EmojiRepository", "Successfully unsaved Emoji (Id: $id)")
+            Result.success(Unit)
+        } else {
+            Log.d("EmojiRepository", "Failed to unsave Emoji (Id: $id), ${response.code()}")
+            Result.failure(Exception("Failed to unsave Emoji (Id: $id), ${response.code()}"))
         }
     }
 
     override suspend fun deleteEmoji(id: String): Response<Unit> {
         TODO("Not yet implemented")
-    }
-
-    fun createVideoThumbnail(context: Context, videoFile: File): File? {
-        val retriever = MediaMetadataRetriever()
-        try {
-            retriever.setDataSource(videoFile.absolutePath)
-            val bitmap = retriever.frameAtTime
-
-            bitmap?.let {
-                val thumbnailFile = File(context.cacheDir, "thumbnail_${videoFile.name}.jpg")
-                FileOutputStream(thumbnailFile).use { out ->
-                    bitmap.compress(Bitmap.CompressFormat.JPEG, 75, out)
-                }
-                Log.d("create_TN", "Thumbnail created: ${thumbnailFile.absolutePath}")
-                return thumbnailFile
-            }
-        } catch (e: Exception) {
-            Log.e("EmojiRepository_create_TN", "ERROR creating thumbnail: ${e.message?:"Unknown error"}")
-        } finally {
-            retriever.release()
-        }
-        return null
     }
 }
